Refactor folder strucutre of tests
