+++ components/cellery-component-test/src/test/java/org/cellery/components/test/scenarios/dockersource/DockerSourceTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.cellery.components.test;
+package org.cellery.components.test.scenarios.dockersource;
 
 import io.cellery.models.Cell;
 import org.ballerinax.kubernetes.exceptions.KubernetesPluginException;
